refactor(cli-validation): separate cli behaviour from validation
This separation will facilitate further refactoring, moving the main validation to lib, split in
multiple files etc.
